Renamed module with view factor methods
